Initial Commit
Abstract 

This project presents a multi-linear regression model used to predict the quality of a white wine. Quality of the wine is the response variable. The goal is to predict the quality of wine using a model which is a good fit for the data by eliminating non-significant attributes. This has been achieved using regression analysis on multiple models using a data set for white wine. A starting model including 11 explanatory variables is considered and regression analysis is performed to obtain better models by eliminating certain predictor variables. Assumptions for each model will be checked and explanatory variables are continually reduced till a model that is a good fit for the data is found. After conducting regression analysis on multiple model, it is found that Alcohol is the most impactful explanatory variable along with 6 other explanatory variables to obtain the best regression model to predict the wine quality. With this model we hope to give novice wine tasters the ability to predict quality of wine without having to taste it.